Mejoras Generales

+++ app/src/main/java/com/coppel/rhconecta/dev/business/models/ConsultaAbonoResponse.java
@@ -25,22 +25,22 @@ public class ConsultaAbonoResponse extends WithDrawSavingBaseResponse {
     }
 
     public class Response {
-        private int imp_cuentacorriente;
+        private Float imp_cuentacorriente;
         private int clv_retirocuentacorriente;
-        private int imp_ahorroadiciona;
+        private Float imp_ahorroadiciona;
         private int clv_retiroahorroadicional;
-        private int imp_fondotrabajador;
+        private Float imp_fondotrabajador;
         private int clv_retirofondotrabajador;
         private String des_proceso;
         private String des_cambiar;
         private String des_mensaje;
         private String des_abonoProceso;
 
-        public int getImp_cuentacorriente() {
+        public Float getImp_cuentacorriente() {
             return imp_cuentacorriente;
         }
 
-        public void setImp_cuentacorriente(int imp_cuentacorriente) {
+        public void setImp_cuentacorriente(Float imp_cuentacorriente) {
             this.imp_cuentacorriente = imp_cuentacorriente;
         }
 
@@ -52,11 +52,11 @@ public class ConsultaAbonoResponse extends WithDrawSavingBaseResponse {
             this.clv_retirocuentacorriente = clv_retirocuentacorriente;
         }
 
-        public int getImp_ahorroadicional() {
+        public Float getImp_ahorroadicional() {
             return imp_ahorroadiciona;
         }
 
-        public void setImp_ahorroadicional(int imp_ahorroadicional) {
+        public void setImp_ahorroadicional(Float imp_ahorroadicional) {
             this.imp_ahorroadiciona = imp_ahorroadicional;
         }
 
@@ -68,11 +68,11 @@ public class ConsultaAbonoResponse extends WithDrawSavingBaseResponse {
             this.clv_retiroahorroadicional = clv_retiroahorroadicional;
         }
 
-        public int getImp_fondotrabajador() {
+        public Float getImp_fondotrabajador() {
             return imp_fondotrabajador;
         }
 
-        public void setImp_fondotrabajador(int imp_fondotrabajador) {
+        public void setImp_fondotrabajador(Float imp_fondotrabajador) {
             this.imp_fondotrabajador = imp_fondotrabajador;
         }
